Updates chart functions.
